- Changed reference to AutoMapper to the range `[11.0, 12.0)` As a breaking change from version `11` onwards is dropped support for `net461` it seems unreasonable to keep support for this framework in the extension also.
- Switched targets to only support `netstandard2.1`
This is due to the fact that AutoMapper from version `11` has dropped support for `net461` target.

Also removed optional dependencies, as package only target single framework now.

- Switched targets for test project to `net6.0`
Mainly because AutoMapper `11` targets `netstandard2.1` and this is not supported in `netcoreapp2.1`, also this target framework is no longer supported generally so updating seems like a good idea.

Secondly I work on an Apple Silicon machine, so bumping to something compatible with .Net 6 was required for me to be able to run the tests.

- Updated using statements to target `AutoMapper.Internal` when necessary
- Use new `.Internal()` rather than old `.Advanced`